Moved RoomCoordinatorAgent's behaviours from class to separate files for clarity, changed args and added usage to the agent, now it's using DF to get the neighbours' and upkeeper's AIDs; fixed typos in README.md
